chain,graph,store: added code to properly resume from a subgraph block ptr and no firehose cursor
Firehose is now using the subgraph ptr if no cursor is found. To ensure we do not create inconsistency in the data due to how Firehose resolves current block number, a sanity check has now been added. This check ensure that the first received block's parent from firehose is the same as the subgraph pointer block. If it's not the case, we revert to the last know final block so we are sure to start from a correct chain segment.

What final block to pick is decided by the chain itself. On Ethereum, we use `block.number - reorg_threshold` (guarded) while on NEAR, we use the last final block height defined by the block itself.

#### Missing/To Change

I pushed the PR to get an overall feeling of the idea and get early inputs.

Right now I'm using `revert_block_operations` but this is wrong as it's able to only revert a single block. I need in fact probably to `rewind` the subgraph but I would like to get advise here on how to do that properly and if there is any risks or stuff I missed with the approach.